refactor(ui): remove unused 'installed' field from agent selection
The 'installed' field was never actually implemented for agent detection.
This commit:
- Removes 'installed' and 'installable' fields from AgenticToolOption type
- Consolidates AgenticToolOption type definition in src/types/ui.ts
- Removes duplicate type definitions across components
- Removes unused onInstall handlers and UI elements
- Fixes type assertion for agentic_tool in ScheduleTab

All agents are now treated as available by default. This simplifies
the UI and removes dead code for a feature that was never built.
