+++ apps/agor-ui/src/components/AgentSelectionCard/AgentSelectionCard.tsx
@@ -1,43 +1,26 @@
-import { CheckCircleOutlined, DownloadOutlined } from '@ant-design/icons';
-import { Button, Card, Space, Tag, Typography } from 'antd';
+import { Card, Space, Typography } from 'antd';
+import type { AgenticToolOption } from '../../types';
 import { ToolIcon } from '../ToolIcon';
-
-// UI-only type for agent selection (different from AgenticTool which has UUIDv7 ID)
-interface AgenticToolOption {
-  id: string; // AgenticToolName as string
-  name: string;
-  icon: string;
-  installed?: boolean;
-  installable?: boolean;
-  version?: string;
-  description?: string;
-}
 
 export interface AgentSelectionCardProps {
   agent: AgenticToolOption;
   selected?: boolean;
   onClick?: () => void;
-  onInstall?: () => void;
 }
 
 export const AgentSelectionCard: React.FC<AgentSelectionCardProps> = ({
   agent,
   selected = false,
   onClick,
-  onInstall,
 }) => {
-  // Treat agents as available by default unless explicitly marked as not installed
-  const isDisabled = agent.installed === false;
-
   return (
     <Card
-      hoverable={!isDisabled}
-      onClick={isDisabled ? undefined : onClick}
+      hoverable
+      onClick={onClick}
       style={{
         borderColor: selected ? '#1890ff' : undefined,
         borderWidth: selected ? 2 : 1,
-        cursor: isDisabled ? 'not-allowed' : 'pointer',
-        opacity: isDisabled ? 0.6 : 1,
+        cursor: 'pointer',
       }}
       styles={{
         body: { padding: 12 },
@@ -50,25 +33,7 @@ export const AgentSelectionCard: React.FC<AgentSelectionCardProps> = ({
             <Typography.Text strong style={{ fontSize: '14px' }}>
               {agent.name}
             </Typography.Text>
-            {agent.installed === false && (
-              <Tag color="orange" style={{ fontSize: '11px', padding: '0 6px' }}>
-                COMING SOON
-              </Tag>
-            )}
           </Space>
-          {!agent.installed && agent.installable && (
-            <Button
-              type="text"
-              size="small"
-              icon={<DownloadOutlined />}
-              onClick={e => {
-                e.stopPropagation();
-                onInstall?.();
-              }}
-            >
-              Install
-            </Button>
-          )}
         </Space>
 
         {agent.version && (
